Complete CheckersBoard __init__
Complete __init__ function in CheckersBoard class. This function
contains the dimensions of the board, types of pieces on the board and
the representation of the board. The board is also initiated, and all
the player pieces are added to their appropriate starting positions.
The board is represented in a 2D list.