Fix computer player's move searching for easy mode
The _get_easy_move() function had been built with the idea that
our game would make multiple moves with one move object but later
was changed to move one tile at a time. This modification to the
easy mode finds a random move or jump in the two front diagonals
and returns a random choice from the possible jumps. If no jumps
exist then return a random regular move.