Add extending the game section to README
Adds the extending the game section which lists some features that were
out of scope and gives an example of extending difficulties for the
computer players.

Note: I had this completed but was asked to work on some code. In the
meantime a member accidently worked on this without noticing the work
breakdown that had been posted previously to the group chat.